Add: Sign personal message

+++ src/parse.h
@@ -92,7 +92,7 @@ typedef struct transactionContext_t {
     uint8_t pathLength;
     uint32_t bip32Path[MAX_BIP32_PATH];
     uint8_t rawTx[MAX_RAW_TX];
-    uint16_t rawTxLength;
+    uint32_t rawTxLength;
     uint8_t hash[HASH_SIZE];
     uint8_t signature[MAX_RAW_SIGNATURE];
     uint8_t signatureLength;
